améliorations méthodes pour gérer les ratings

+++ Poseiden-skeleton/src/main/java/com/poseidon/api/repositories/RatingRepository.java
@@ -9,6 +9,6 @@ import java.util.Optional;
 @Repository
 public interface RatingRepository extends JpaRepository<Rating, Long> {
 
-    Optional<Rating> findRatingById(Integer id);
+    Optional<Rating> findRatingById(Long id);
 
 }
